+++ ui/app/settings.js
@@ -1,7 +1,7 @@
 const { Component } = require('react')
 const PropTypes = require('prop-types')
 const h = require('react-hyperscript')
-const { connect } = require('react-redux')
+const connect = require('react-redux').connect
 const actions = require('./actions')
 const infuraCurrencies = require('./infura-conversion.json')
 const validUrl = require('valid-url')
@@ -9,8 +9,8 @@ const { exportAsFile } = require('./util')
 const TabBar = require('./components/tab-bar')
 const SimpleDropdown = require('./components/dropdowns/simple-dropdown')
 const ToggleButton = require('react-toggle-button')
+const locales = require('../../app/_locales/index.json')
 const { OLD_UI_NETWORK_TYPE } = require('../../app/scripts/config').enums
-const t = require('../i18n')
 
 const getInfuraCurrencyOptions = () => {
   const sortedCurrencies = infuraCurrencies.objects.sort((a, b) => {
@@ -22,6 +22,16 @@ const getInfuraCurrencyOptions = () => {
       displayValue: `${code.toUpperCase()} - ${name}`,
       key: code,
       value: code,
+    }
+  })
+}
+
+const getLocaleOptions = () => {
+  return locales.map((locale) => {
+    return {
+      displayValue: `${locale.name}`,
+      key: locale.code,
+      value: locale.code,
     }
   })
 }
@@ -45,8 +55,8 @@ class Settings extends Component {
     return h('div.settings__tabs', [
       h(TabBar, {
         tabs: [
-          { content: t('settings'), key: 'settings' },
-          { content: t('info'), key: 'info' },
+          { content: this.context.t('settings'), key: 'settings' },
+          { content: this.context.t('info'), key: 'info' },
         ],
         defaultTab: activeTab,
         tabSelected: key => this.setState({ activeTab: key }),
@@ -59,7 +69,7 @@ class Settings extends Component {
 
     return h('div.settings__content-row', [
       h('div.settings__content-item', [
-        h('span', t('blockiesIdenticon')),
+        h('span', this.context.t('blockiesIdenticon')),
       ]),
       h('div.settings__content-item', [
         h('div.settings__content-item-col', [
@@ -79,16 +89,41 @@ class Settings extends Component {
 
     return h('div.settings__content-row', [
       h('div.settings__content-item', [
-        h('span', t('currentConversion')),
+        h('span', this.context.t('currentConversion')),
         h('span.settings__content-description', `Updated ${Date(conversionDate)}`),
       ]),
       h('div.settings__content-item', [
         h('div.settings__content-item-col', [
           h(SimpleDropdown, {
-            placeholder: t('selectCurrency'),
+            placeholder: this.context.t('selectCurrency'),
             options: getInfuraCurrencyOptions(),
             selectedOption: currentCurrency,
             onSelect: newCurrency => setCurrentCurrency(newCurrency),
+          }),
+        ]),
+      ]),
+    ])
+  }
+
+  renderCurrentLocale () {
+    const { updateCurrentLocale, currentLocale } = this.props
+    const currentLocaleMeta = locales.find(locale => locale.code === currentLocale)
+    const currentLocaleName = currentLocaleMeta ? currentLocaleMeta.name : ''
+
+    return h('div.settings__content-row', [
+      h('div.settings__content-item', [
+        h('span', 'Current Language'),
+        h('span.settings__content-description', `${currentLocaleName}`),
+      ]),
+      h('div.settings__content-item', [
+        h('div.settings__content-item-col', [
+          h(SimpleDropdown, {
+            placeholder: 'Select Locale',
+            options: getLocaleOptions(),
+            selectedOption: currentLocale,
+            onSelect: async (newLocale) => {
+              updateCurrentLocale(newLocale)
+            },
           }),
         ]),
       ]),
@@ -102,16 +137,17 @@ class Settings extends Component {
     switch (provider.type) {
 
       case 'mainnet':
-        title = t('currentNetwork')
-        value = t('mainnet')
+        title = this.context.t('currentNetwork')
+        value = this.context.t('mainnet')
         color = '#038789'
         break
+
 
  
 
 
       default:
-        title = t('currentRpc')
+        title = this.context.t('currentRpc')
         value = provider.rpcTarget
     }
 
@@ -132,12 +168,12 @@ class Settings extends Component {
     return (
       h('div.settings__content-row', [
         h('div.settings__content-item', [
-          h('span', t('newRPC')),
+          h('span', this.context.t('newRPC')),
         ]),
         h('div.settings__content-item', [
           h('div.settings__content-item-col', [
             h('input.settings__input', {
-              placeholder: t('newRPC'),
+              placeholder: this.context.t('newRPC'),
               onChange: event => this.setState({ newRpc: event.target.value }),
               onKeyPress: event => {
                 if (event.key === 'Enter') {
@@ -150,7 +186,7 @@ class Settings extends Component {
                 event.preventDefault()
                 this.validateRpc(this.state.newRpc)
               },
-            }, t('save')),
+            }, this.context.t('save')),
           ]),
         ]),
       ])
@@ -166,9 +202,9 @@ class Settings extends Component {
       const appendedRpc = `http://${newRpc}`
 
       if (validUrl.isWebUri(appendedRpc)) {
-        displayWarning(t('uriErrorMsg'))
+        displayWarning(this.context.t('uriErrorMsg'))
       } else {
-        displayWarning(t('invalidRPC'))
+        displayWarning(this.context.t('invalidRPC'))
       }
     }
   }
@@ -177,10 +213,10 @@ class Settings extends Component {
     return (
       h('div.settings__content-row', [
         h('div.settings__content-item', [
-          h('div', t('stateLogs')),
+          h('div', this.context.t('stateLogs')),
           h(
             'div.settings__content-description',
-            t('stateLogsDescription')
+            this.context.t('stateLogsDescription')
           ),
         ]),
         h('div.settings__content-item', [
@@ -189,13 +225,13 @@ class Settings extends Component {
               onClick (event) {
                 window.logStateString((err, result) => {
                   if (err) {
-                    this.state.dispatch(actions.displayWarning(t('stateLogError')))
+                    this.state.dispatch(actions.displayWarning(this.context.t('stateLogError')))
                   } else {
                     exportAsFile('MetaMask State Logs.json', result)
                   }
                 })
               },
-            }, t('downloadStateLogs')),
+            }, this.context.t('downloadStateLogs')),
           ]),
         ]),
       ])
@@ -207,7 +243,7 @@ class Settings extends Component {
 
     return (
       h('div.settings__content-row', [
-        h('div.settings__content-item', t('revealSeedWords')),
+        h('div.settings__content-item', this.context.t('revealSeedWords')),
         h('div.settings__content-item', [
           h('div.settings__content-item-col', [
             h('button.btn-primary--lg.settings__button--red', {
@@ -215,7 +251,7 @@ class Settings extends Component {
                 event.preventDefault()
                 revealSeedConfirmation()
               },
-            }, t('revealSeedWords')),
+            }, this.context.t('revealSeedWords')),
           ]),
         ]),
       ])
@@ -227,7 +263,7 @@ class Settings extends Component {
 
     return (
       h('div.settings__content-row', [
-        h('div.settings__content-item', t('useOldUI')),
+        h('div.settings__content-item', this.context.t('useOldUI')),
         h('div.settings__content-item', [
           h('div.settings__content-item-col', [
             h('button.btn-primary--lg.settings__button--orange', {
@@ -235,7 +271,7 @@ class Settings extends Component {
                 event.preventDefault()
                 setFeatureFlagToBeta()
               },
-            }, t('useOldUI')),
+            }, this.context.t('useOldUI')),
           ]),
         ]),
       ])
@@ -246,7 +282,7 @@ class Settings extends Component {
     const { showResetAccountConfirmationModal } = this.props
 
     return h('div.settings__content-row', [
-      h('div.settings__content-item', t('resetAccount')),
+      h('div.settings__content-item', this.context.t('resetAccount')),
       h('div.settings__content-item', [
         h('div.settings__content-item-col', [
           h('button.btn-primary--lg.settings__button--orange', {
@@ -254,7 +290,7 @@ class Settings extends Component {
               event.preventDefault()
               showResetAccountConfirmationModal()
             },
-          }, t('resetAccount')),
+          }, this.context.t('resetAccount')),
         ]),
       ]),
     ])
@@ -267,6 +303,7 @@ class Settings extends Component {
       h('div.settings__content', [
         warning && h('div.settings__error', warning),
         this.renderCurrentConversion(),
+        this.renderCurrentLocale(),
         // this.renderCurrentProvider(),
         this.renderNewRpcUrl(),
         this.renderStateLogs(),
@@ -290,7 +327,7 @@ class Settings extends Component {
   renderInfoLinks () {
     return (
       h('div.settings__content-item.settings__content-item--without-height', [
-        h('div.settings__info-link-header', t('links')),
+        h('div.settings__info-link-header', this.context.t('links')),
         h('div.settings__info-link-item', [
           h('a', {
             href: 'https://nekonium.github.io/',
@@ -335,7 +372,7 @@ class Settings extends Component {
             h('div.settings__info-item', [
               h(
                 'div.settings__info-about',
-                t('builtInCalifornia')
+                this.context.t('builtInCalifornia')
               ),
             ]),
             h('div.settings__info-item', [
@@ -388,6 +425,9 @@ Settings.propTypes = {
   warning: PropTypes.string,
   goHome: PropTypes.func,
   isMascara: PropTypes.bool,
+  updateCurrentLocale: PropTypes.func,
+  currentLocale: PropTypes.string,
+  t: PropTypes.func,
 }
 
 const mapStateToProps = state => {
@@ -395,6 +435,7 @@ const mapStateToProps = state => {
     metamask: state.metamask,
     warning: state.appState.warning,
     isMascara: state.metamask.isMascara,
+    currentLocale: state.metamask.currentLocale,
   }
 }
 
@@ -406,6 +447,7 @@ const mapDispatchToProps = dispatch => {
     displayWarning: warning => dispatch(actions.displayWarning(warning)),
     revealSeedConfirmation: () => dispatch(actions.revealSeedConfirmation()),
     setUseBlockie: value => dispatch(actions.setUseBlockie(value)),
+    updateCurrentLocale: key => dispatch(actions.updateCurrentLocale(key)),
     setFeatureFlagToBeta: () => {
       return dispatch(actions.setFeatureFlag('betaUI', false, 'OLD_UI_NOTIFICATION_MODAL'))
         .then(() => dispatch(actions.setNetworkEndpoints(OLD_UI_NETWORK_TYPE)))
@@ -414,6 +456,10 @@ const mapDispatchToProps = dispatch => {
       return dispatch(actions.showModal({ name: 'CONFIRM_RESET_ACCOUNT' }))
     },
   }
+}
+
+Settings.contextTypes = {
+  t: PropTypes.func,
 }
 
 module.exports = connect(mapStateToProps, mapDispatchToProps)(Settings)