+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Current Master
 
+## 4.5.0 Mon Apr 02 2018
+
+- (beta ui) Internationalization: Select your preferred language in the settings screen
+- Internationalization: various locale improvements
+- Fix bug where the "Reset account" feature would not clear the network cache.
+- Increase maximum gas limit, to allow very gas heavy transactions, since block gas limits have been stable.
+
 ## 4.4.0 Mon Mar 26 2018
 
 - Internationalization: Taiwanese, Thai, Slovenian